+++ src/pipelines/onnx/models.rs
@@ -8,6 +8,9 @@ use crate::pipelines::onnx::decoder::ONNXDecoder;
 use crate::pipelines::onnx::encoder::ONNXEncoder;
 use crate::{Config, RustBertError};
 
+use crate::pipelines::onnx::conversion;
+use ndarray::IxDyn;
+use ort::tensor::DynOrtTensor;
 use ort::Environment;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -1093,3 +1096,29 @@ impl PrivateLanguageGenerator for ONNXConditionalGenerator {
 }
 
 impl LanguageGenerator for ONNXConditionalGenerator {}
+
+#[derive(Debug)]
+/// Container used to store key-value cached states for efficient decoding.
+pub struct ONNXLayerCache {
+    pub values: HashMap<String, Tensor>,
+}
+
+impl ONNXLayerCache {
+    /// Helper function to create a cache layer from an ONNX model output.
+    /// Assumes that the output names for cached keys and values contain `key` and `value` in their name, respectively.
+    pub fn from_ort_output(
+        ort_output: &[DynOrtTensor<IxDyn>],
+        key_value_names: &HashMap<String, usize>,
+    ) -> Result<ONNXLayerCache, RustBertError> {
+        let values = key_value_names
+            .iter()
+            .filter(|(name, _)| name.contains("key") | name.contains("value"))
+            .map(|(name, pos)| {
+                let value = &ort_output[*pos];
+                Ok((name.to_string(), conversion::ort_tensor_to_tch(value)?))
+            })
+            .collect::<Result<HashMap<String, Tensor>, RustBertError>>()?;
+
+        Ok(ONNXLayerCache { values })
+    }
+}